Modularize Weather pipeline
